overhaul of the methods. made them into doubles instead of int to account for decimals.

+++ allshapes.cpp
@@ -9,12 +9,12 @@ circle :: circle(int radius)
   this->radius = radius;
 }
 
-int circle :: getArea()
+double circle :: getArea()
 {
     return M_PI * pow(this->radius, 2.0);
 };
 
-int circle :: getPerimeter()
+double circle :: getPerimeter()
 {
     return 2 * M_PI * this->radius;
 };
@@ -28,12 +28,12 @@ rectangle :: rectangle(int length, int width)
   this->width = width;
 }
 
-int rectangle :: getArea()
+double rectangle :: getArea()
 {
     return this->length * this->width;
 };
 
-int rectangle :: getPerimeter()
+double rectangle :: getPerimeter()
 {
     return 2*(this->length+this->width);
 };
@@ -47,12 +47,12 @@ rightTriangle :: rightTriangle(int leg1, int leg2)
   this->leg2 = leg2;
 }
 
-int rightTriangle :: getArea()
+double rightTriangle :: getArea()
 {
     return (this->leg1*this->leg2)/2;
 };
 
-int rightTriangle :: getPerimeter()
+double rightTriangle :: getPerimeter()
 {
     return this->leg1+this->leg2+(sqrt(pow(this->leg1, 2.0)+pow(this->leg2, 2.0)));
 };
