+++ allshapes.hpp
@@ -7,8 +7,8 @@ private:
 
 public:
    explicit circle(int radius);
-   int getArea() override;
-   int getPerimeter() override;
+   double getArea() override;
+   double getPerimeter() override;
 
 };
 
@@ -20,8 +20,8 @@ private:
 
 public: 
    explicit rectangle (int length, int width);
-   int getArea() override;
-   int getPerimeter() override;
+   double getArea() override;
+   double getPerimeter() override;
 
 };
 
@@ -34,7 +34,7 @@ private:
 
 public: 
    explicit rightTriangle(int leg1, int leg2);
-   int getArea() override;
-   int getPerimeter() override;
+   double getArea() override;
+   double getPerimeter() override;
 
 };
