+++ shape.hpp
@@ -5,8 +5,8 @@ class shape {
 
 public:
 
-   virtual int getArea() = 0;
-   virtual int getPerimeter() = 0;
+   virtual double getArea() = 0;
+   virtual double getPerimeter() = 0;
 
 };
 #endif
